% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstDIF-class_methods.R
\name{mstDIF-Methods}
\alias{mstDIF-Methods}
\title{Methods for the mstDIF-class}
\arguments{
\item{x}{an object of class \code{mstDIF}}

\item{object}{an object of class \code{mstDIF}}

\item{item_selection}{either \code{NULL} or an integer vector selecting the
item numbers. When \code{items = NULL} (the default), the DIF test
is done for all items.}

\item{DIF_type}{a string that should one or more of "overall", "uniform",
"non-uniform", "all".}

\item{ordered}{logical: should the summary be ordered according to the obtained p-values (in ascending order)?}

\item{...}{other argements passed to the method.}
}
\description{
\code{print} and \code{summary}methods for objects of the
\code{mstDIF-class}, as returned by \code{\link{mstDIF}}. See details
for more information about the methods.
}
\details{
The \code{print} method prints some basic information about the
  \code{mstDIF-class} object.

The \code{summary} method computes a data frame with a row for each item
that was included in the test. The columns are:
  \describe{
     \item{item}{The name of the item}
     \item{statistic}{The value for the used statistic per item}
     \item{p_value}{The p-value per item}
     \item{eff_size}{An effect-size for the DIF-test, if applicable}
   }
}
