% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstDIF.R
\name{mstDIF}
\alias{mstDIF}
\title{A general function to detect differential item functioning (DIF) in multistage tests (MSTs)}
\usage{
mstDIF(resp = NULL, mirt_object = NULL, order_by, method, ...)
}
\arguments{
\item{resp}{A data frame containing the response matrix. Rows correspond to respondents, columns to items.
Not required for analytical score-based tests.}

\item{mirt_object}{A SingleGroup or Multigroup object as it is returned by mirt. Only required for
analytical score-based tests.}

\item{order_by}{A vector of ability estimates for each respondent.}

\item{method}{A character value indicating the DIF test that should be used. Possible values are "logreg"
(Logistic regression), "mstsib" (mstSIB), "bootstrap" (score-based Bootstrap test), "permutation" (score-based)
permutation test) and "analytical" (analytical score-based test)}

\item{...}{additional, test-specific arguments}
}
\value{
A list with the following elements:
\describe{
  \item{\code{p_values}}{A data frame where each row corresponds to an item. The first column corresponds to the
   item numbers, the second column to the p-value of an overall DIF test for ths corresponding item.}
  \item{\code{resp}}{The response matrix as a data frame.}
  \item{\code{order_by}}{The person covariate tested for DIF.}
  \item{\code{method}}{The selected DIF test.}
  \item{\code{complete_output}}{The complete output of the selected DIF test.
  Differs depending on the selection.}
}
}
\description{
This function allows the application of various methods for the detection of differential item functioning
in multistage tests. Currently five methods are implemented: 1. Logistic Regression, 2. mstSIB, 3. analytical
score-base tests, 4. a score-based Bootstrap test, 5. a score-based permutation test. The required input
differs with regard to the selected DIF test.
}
\details{
Author: Rudolf Debelak
}
