% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstDIF-package.R, R/mstDIF.R
\docType{package}
\name{mstDIF}
\alias{mstDIF}
\alias{mstDIF.default}
\alias{mstDIF.AllModelClass}
\alias{mstDIF.dRm}
\title{mstDIF: A Collection of Statistical Tests for DIF Detection in Multistage Tests}
\usage{
mstDIF(resp, object, DIF_covariate, method, theta, see, theta_method, ...)

\method{mstDIF}{default}(
  resp,
  object = NULL,
  DIF_covariate,
  method,
  theta = NULL,
  see = NULL,
  theta_method = NULL,
  ...
)

\method{mstDIF}{AllModelClass}(
  resp = NULL,
  object,
  DIF_covariate,
  method,
  theta = NULL,
  see = NULL,
  theta_method = "WLE",
  ...
)

\method{mstDIF}{dRm}(
  resp = NULL,
  object,
  DIF_covariate,
  method,
  theta = NULL,
  see = NULL,
  theta_method = NULL,
  ...
)
}
\arguments{
\item{resp}{A data frame or matrix containing the response matrix. Rows correspond to respondents, columns to items.}

\item{object}{A SingleGroup-class or MultiGroup-class object as returned by mirt, or a dRm objcect as returned by the RM function in eRm.}

\item{DIF_covariate}{A vector of ability estimates for each respondent.}

\item{method}{A character value indicating the DIF test that should be used. Possible values are "logreg"
(Logistic regression), "mstsib" (mstSIB), "bootstrap" (score-based Bootstrap test), "permutation" (score-based)
permutation test) and "analytical" (analytical score-based test)}

\item{theta}{Estimates of the ability parameters}

\item{see}{Estimates of the standard error of estimation.}

\item{theta_method}{Method for estimating the ability parameters if they
should be estimated based on the responses. The calculation is carried
out by the mirt package. Can be: "WLE" (default),
"MAP", "EAP", "ML", "EAPsum", "plausible", "classify"}

\item{...}{additional, test-specific arguments}
}
\value{
A list with the following elements:
\describe{
  \item{\code{resp}}{The response matrix as a data frame.}
  \item{\code{method}}{The used DIF detection method.}
  \item{\code{test}}{The used test or statistic.}
  \item{\code{DIF_covariate}}{The person covariate tested for DIF.}
  \item{\code{DIF_test}}{A list with the DIF-test results.}
  \item{\code{call}}{The function call.}
  \item{\code{method_results}}{The complete output of the selected DIF test.
  Differs depending on the selection.}
}
}
\description{
This function allows the application of various methods for the detection of differential item functioning
in multistage tests. Currently five methods are implemented: 1. Logistic Regression, 2. mstSIB, 3. analytical
score-base tests, 4. a score-based Bootstrap test, 5. a score-based permutation test. The required input
differs with regard to the selected DIF test.
}
\details{
Author: Rudolf Debelak and Dries Debeer
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default mstDIF method

\item \code{AllModelClass}: mstDIF method for mirt-objects

\item \code{dRm}: mstDIF method for dRm-objects
}}

